LayoutTests/media/video-source-type.html

<!DOCTYPE html>
<html>
<head>
    <title> &lt;source&gt; @type attribute</title>

    <script src=../http/tests/media/media-file.js></script>
    <script src=../http/tests/media/video-test.js></script>
    <script>
        function setup()
        {
            video = mediaElement = document.getElementsByTagName('video')[0];
        
            waitForEvent("loadstart", function () {
                testExpected("stripExtension(relativeURL(video.currentSrc))", stripExtension(findMediaFile("video", "content/test")));
                endTest();
            });
        }
    </script>
</head>
<body>
    <video controls>
        <source src=content/error.mpeg type=video/blahblah>
        <source src=test.mp4 type=video/x-chicken-face>
        <source src=test.ogv type=audio/x-higglety-pigglety>
        <source src=content/test.mp4 type=video/mpeg>
        <source src=content/test.ogv type=video/ogg>
        <source src=content/error2.mpeg type=video/mpeg>
    </video>
    <script>setup();</script>
</body>
</html>
