LayoutTests/media/video-source-type-expected.txt

EVENT(loadstart) TEST(relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
